@@ -1,5 +1,3 @@
-'use strict';
-
 /*
  * Copy of Erik's caller module with node 8 support added.
 */
@@ -10,7 +8,7 @@
  * @blessings: https://twitter.com/eriktoth/statuses/413719312273125377
  * @see https://code.google.com/p/v8/wiki/JavaScriptStackTraceApi
  */
-const caller = function (depth) {
+export const caller = function (depth) {
     const pst = Error.prepareStackTrace;
 
     Error.prepareStackTrace = function (_, frames) {
@@ -34,4 +32,4 @@ const caller = function (depth) {
     return file;
 };
 
-module.exports = caller;
+export default caller;
