@@ -1,16 +1,13 @@
-'use strict';
-
-const Package = require('../package.json');
-const Joi = require('joi');
-const Hoek = require('@hapi/hoek');
-const Caller = require('./caller');
-const Path = require('path');
-const Parser = require('swagger-parser');
-const Utils = require('./utils');
-const Routes = require('./routes');
-const Yaml = require('js-yaml');
-const Fs = require('fs');
-const Util = require('util');
+import Package from '../package.json';
+import Joi from 'joi';
+import Hoek from '@hapi/hoek';
+import Caller from './caller.js';
+import Path from 'path';
+import Parser from 'swagger-parser';
+import Utils from './utils.js';
+import { create } from './routes.js';
+import Yaml from 'js-yaml';
+import Fs from 'fs';
 
 const CALLER_DIR = Path.resolve(Path.dirname(Caller()));
 
@@ -41,7 +38,7 @@ const stripVendorExtensions = function (obj) {
         return clean;
     }
 
-    if (Util.isObject(obj)) {
+    if (typeof obj === 'object') {
         const clean = {};
         for (const [key, value] of Object.entries(obj)) {
             if (!key.match(/\x-(.*)/)) {
@@ -55,7 +52,7 @@ const stripVendorExtensions = function (obj) {
     return obj;
 };
 
-const requireApi = function (path) {
+const requireApi = async function (path) {
     let document;
 
     if (path.match(/\.ya?ml?/)) {
@@ -63,7 +60,8 @@ const requireApi = function (path) {
         document = Yaml.load(file);
     }
     else {
-        document = require(path);
+        document = await import(path);
+        document = Utils.mergeDefault(document);
     }
 
     return document;
@@ -72,7 +70,7 @@ const requireApi = function (path) {
 const register = async function (server, options, next) {
 
     //Validator needs to be explicitly declared for Hapi v19.*
-    server.validator(require('joi'));
+    server.validator(Joi);
 
     const validation = optionsSchema.validate(options);
 
@@ -103,7 +101,7 @@ const register = async function (server, options, next) {
     let apiDocument;
 
     if (typeof api === 'string') {
-        apiDocument = requireApi(api);
+        apiDocument = await requireApi(api);
         basedir = Path.dirname(Path.resolve(api));
     }
     else {
@@ -113,7 +111,8 @@ const register = async function (server, options, next) {
 
     if (spec['x-hapi-auth-schemes']) {
         for (const [name, path] of Object.entries(spec['x-hapi-auth-schemes'])) {
-            const scheme = require(Path.resolve(Path.join(basedir, path)));
+            let scheme = await import(Path.resolve(Path.join(basedir, path)));
+            scheme = Utils.mergeDefault(scheme);
 
             await server.register({
                 plugin: scheme,
@@ -125,7 +124,7 @@ const register = async function (server, options, next) {
     }
 
     let securitySchemes;
-    
+
     if (spec.swagger && spec.securityDefinitions) {
         securitySchemes = spec.securityDefinitions;
     }
@@ -137,7 +136,8 @@ const register = async function (server, options, next) {
     if (securitySchemes) {
         for (const [name, security] of Object.entries(securitySchemes)) {
             if (security['x-hapi-auth-strategy']) {
-                const strategy = require(Path.resolve(Path.join(basedir, security['x-hapi-auth-strategy'])));
+                let strategy = await import(Path.resolve(Path.join(basedir, security['x-hapi-auth-strategy'])));
+                strategy = Utils.mergeDefault(strategy);
 
                 await server.register({
                     plugin: strategy,
@@ -161,7 +161,7 @@ const register = async function (server, options, next) {
     }
 
     let apiPath = docs.path;
-    if (docs.prefixBasePath){
+    if (docs.prefixBasePath) {
         docs.path = Utils.prefix(docs.path, '/');
         docs.path = Utils.unsuffix(docs.path, '/');
         apiPath = spec.basePath + docs.path;
@@ -188,11 +188,11 @@ const register = async function (server, options, next) {
         vhost
     });
 
-    const routes = await Routes.create(server, { api: spec, basedir, cors, vhost, handlers, extensions, outputvalidation });
+    const routes = await create(server, { api: spec, basedir, cors, vhost, handlers, extensions, outputvalidation });
 
     for (const route of routes) {
         server.route(route);
     }
 };
 
-module.exports.plugin = { register, name: 'openapi', version: Package.version, multiple: true };
+export default { plugin: { register, name: 'openapi', version: Package.version, multiple: true } };
