@@ -1,15 +1,13 @@
-'use strict';
-
-const ObjectFiles = require('merge-object-files');
-const Validators = require('./validators');
-const Hoek = require('@hapi/hoek');
-const Utils = require('./utils');
-const Path = require('path');
-const Props = require('dot-prop');
+import ObjectFiles from 'merge-object-files-es6';
+import Validators from './validators.js';
+import Hoek from '@hapi/hoek';
+import Utils from './utils.js';
+import Path from 'path';
+import Props from 'dot-prop';
 
 const SEPARATOR = String.fromCharCode(0x7F);
 
-const create = async function (server, { api, basedir, cors, vhost, handlers, extensions, outputvalidation }) {
+export const create = async function (server, { api, basedir, cors, vhost, handlers, extensions, outputvalidation }) {
     const routes = [];
     const validator = Validators.create({ api });
 
@@ -27,9 +25,9 @@ const create = async function (server, { api, basedir, cors, vhost, handlers, ex
                     handlers = {};
                 }
 
-                const xhandler = require(Path.resolve(Path.join(basedir, operations['x-hapi-handler'])));
+                const xhandler = await import(Path.resolve(Path.join(basedir, operations['x-hapi-handler'])));
 
-                Props.set(handlers, pathnames, xhandler);
+                Props.set(handlers, pathnames, Utils.mergeDefault(xhandler));
             }
         }
     }
@@ -121,7 +119,7 @@ const create = async function (server, { api, basedir, cors, vhost, handlers, ex
                         options.auth.strategies.push(securityDefinitionName);
                     }
                 }
-                
+
                 if (options.auth.access.scope.length === 0) {
                     options.auth.access.scope = false;
                 }
@@ -138,5 +136,3 @@ const create = async function (server, { api, basedir, cors, vhost, handlers, ex
 
     return routes;
 };
-
-module.exports = { create };
