@@ -1,6 +1,4 @@
-'use strict';
-
-const utils = {
+export const utils = {
 
     verbs: [
         'get',
@@ -80,7 +78,24 @@ const utils = {
         }
 
         return str;
+    },
+
+    mergeDefault: (obj) => {
+        const merged_obj = {};
+        if (obj.default) {
+            for (const [k, v] of Object.entries(obj.default)) {
+                merged_obj[k] = v;
+            }
+        }
+
+        for (const [k, v] of Object.entries(obj)) {
+            if (k !== 'default') {
+                merged_obj[k] = v;
+            }
+        }
+
+        return merged_obj;
     }
 };
 
-module.exports = utils;
+export default utils;
