@@ -1,7 +1,5 @@
-'use strict';
-
-const Enjoi = require('enjoi');
-const Joi = require('joi');
+import Enjoi from 'enjoi';
+import Joi from 'joi';
 
 const extensions = [
     { type: 'int64', base: Joi.string().regex(/^\d+$/) },
@@ -46,7 +44,7 @@ const refineSchema = function (joiSchema, jsonSchema) {
 
 const enjoi = Enjoi.defaults({ extensions, refineType, refineSchema });
 
-const create = function (options = {}) {
+export const create = function (options = {}) {
     const makeValidator = function (parameter, consumes, openapi, allowUnknownProperties = false) {
         const coerce = coercion(parameter, consumes);
 
@@ -83,7 +81,7 @@ const create = function (options = {}) {
 
             if (openapi) {
                 template = { ...template, ...parameter.schema };
-            } 
+            }
             else {
                 template.schema = parameter.schema;
             }
@@ -153,7 +151,7 @@ const create = function (options = {}) {
                             break;
                         }
                     }
-                } 
+                }
                 else {
                     schemaDesc = response.schema;
                 }
@@ -302,7 +300,7 @@ const coercion = function (parameter, consumes) {
                 if (parameter.format === 'int64') {
                     return data;
                 }
-                
+
                 return Number(data);
             };
 
@@ -325,7 +323,7 @@ const coercion = function (parameter, consumes) {
                     in: parameter.in
                 };
             };
-            
+
             break;
         }
     }
@@ -337,4 +335,4 @@ const coercion = function (parameter, consumes) {
     return fn;
 };
 
-module.exports = { create };
+export default { create };
